refactor: lowercase names for folders and files
